pyproject.toml
[tool.poetry]
name = "django-project-default"
version = "0.1.0"
description = ""
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
Django = "^3.1.6"
python-dotenv = "^0.15.0"
djangorestframework = "^3.12.2"
djangorestframework-jwt = "^1.11.0"
django-filter = "^2.4.0"
django-compressor = "^2.4"
channels-redis = "^3.2.0"
channels = "^3.0.3"
Markdown = "^3.3.3"
django-cors-headers = "^3.7.0"
django-sass-processor = "^0.8.2"
slackweb = "^1.0.5"
numpy = "^1.20.1"
mysqlclient = "^2.0.3"
requests-oauthlib = "^1.3.0"
drf-yasg = "^1.20.0"
google-cloud-translate = "1.7.0"
drf-nested-routers = "^0.93.3"
pandas = "^1.5.1"
attrs = "^22.1.0"
django-ipware = "^4.0.2"
boto3 = "^1.26.3"
django-ses = "^3.2.2"
django-storages = "^1.13.1"
bleach = "^5.0.1"
firebase-admin = "v5.2.0"
openai = "^0.25.0"
pillow = "^10.4.0"
coreapi = "^2.3.3"

[poetry.group.dev.dependencies]
httpie = "^2.4.0"
flake8 = "^3.8.4"
yapf = "^0.30.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
